+++ src/Event/DeleteEvent.php
@@ -1,0 +1,31 @@
+<?php
+
+namespace TK\GitHubWebhook\Event;
+
+use TK\GitHubWebhook\Model\Delete\RefType;
+use TK\GitHubWebhook\Model\InstallationLite;
+use TK\GitHubWebhook\Model\Repository;
+use TK\GitHubWebhook\Model\User;
+use TK\GitHubWebhook\Util;
+
+class DeleteEvent extends AbstractEvent
+{
+    public string $ref;
+    public RefType $ref_type;
+    public string $pusher_type;
+    public InstallationLite|null $installation;
+
+    public static  function fromArray(array $data): DeleteEvent
+    {
+        $repository = Util::getArgSafe($data, "repository", Repository::fromArray(...));
+        $sender = Util::getArgSafe($data, "sender", User::fromArray(...));
+        $organization = Util::getArgSafe($data, "organization", User::fromArray(...));
+
+        $instance = new DeleteEvent($repository, $sender, $organization);
+        $instance->ref = $data["ref"];
+        $instance->ref_type = RefType::from($data["ref_type"]);
+        $instance->pusher_type = $data["pusher_type"];
+        $instance->installation = Util::getArgSafe($data, "installation", InstallationLite::fromArray(...));
+        return $instance;
+    }
+}