+++ src/Model/Delete/RefType.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace TK\GitHubWebhook\Model\Delete;
+
+enum RefType: string
+{
+    case TAG = "tag";
+    case BRANCH = "branch";
+}